+++ NEWS.md
@@ -3,13 +3,32 @@
 
 ## 1.8.1
 
-**FFTrees** version 1.8.1 was released [on CRAN](https://CRAN.R-project.org/package=FFTrees) [on 2023-01-10]. 
+**FFTrees** version 1.8.1 is to be released [on CRAN](https://CRAN.R-project.org/package=FFTrees) [on 2023-01-12]. 
 This version mostly fixes an error in a package vignette. 
 
 <!-- Log of changes: --> 
 
 Changes since last release: 
 
+<!-- Blank line. --> 
+
+### Major changes 
+
+- None yet. 
+
+<!-- Blank line. --> 
+
+
+### Minor changes 
+
+- None yet. 
+
+<!-- Blank line. --> 
+
+
+### Details 
+
+- Added check that criterion variable occurs in data.
 - Fixed CRAN error (in `mushrooms.fft` of "Examples of FFTrees"). 
 - Replaced Travis badge in `README.Rmd` by GitHub action on R-CMD-check (standard). 
 - Minor revision of vignettes.
@@ -378,6 +397,6 @@ Thus, the main tree building function is now `FFTrees()` and the new tree object
 
 ------ 
 
-[File `NEWS.md` last updated on 2023-01-09.]
+[File `NEWS.md` last updated on 2023-01-12.]
 
 <!-- eof. -->